Refactor wildfly cli handler
